AG-18493 Extract the partial resolve cache from OptionsGraph
optionsGraph.ts is dense enough that anything not about the graph itself makes
it harder to follow. Move the cache added for per-datum styler resolution into
optionsPartialCache.ts, alongside the existing optionsStructuralCache, leaving
the graph with only the wiring: the field, two invalidation points, and the
lookup-or-resolve block in resolvePartialInternal. Its share of the change
drops from 137 added lines to 33, ten of which are the split-out
resolvePartialUncached signature.

Key building, the copy discipline and the CSS-variable comparison move with
it, so the reason results are copied on the way in and out — line and area
series set `marker ??= {}` on what they are handed — is stated next to the
code that copies rather than inferred at the call site.

Two details are not a straight move:

- LRUCache from ag-charts-core replaces the hand-rolled size cap. A chart with
  more distinct styles than the limit now keeps the most recently used rather
  than whichever arrived first.
- Reads return the entry in a wrapper, so one lookup separates a cached
  `undefined` from a miss instead of needing both `get` and `has`.

No behavioural change otherwise. Creating a 1000 marker quadrant chart takes
208.0ms against 205.3ms before the move, and 317.7ms with no cache at all.
